トークン認証の仕組みを追加

+++ main_app.py
@@ -1,7 +1,8 @@
-from fastapi import FastAPI
+from fastapi import FastAPI, Header, Depends, HTTPException
 import auth
 import item
 from post_parameter import PostCredential, PostRefreshToken, PostItem
+import datetime
 
 tags_metadata = [
     {
@@ -24,6 +25,21 @@ app = FastAPI(
     openapi_tags=tags_metadata
 )
 
+
+def get_auth_token(Authorization: str = Header(None)):
+    if Authorization is None:
+        raise HTTPException(
+            401, headers={"WWW-Authenticate": 'Bearer realm="token_required"'})
+    if Authorization[:7] != "Bearer ":
+        raise HTTPException(
+            400, headers={"WWW-Authenticate": 'Bearer realm="invalid_request"'})
+    return Authorization[7:]
+
+
+def is_authorized(Authorization: str = Header(None)):
+    token = get_auth_token(Authorization)
+    return auth_repository.isInclude(token, datetime.datetime.now())
+
 # ------------------------------------------------------------------------------
 
 
@@ -36,6 +52,11 @@ def create_token(credential: PostCredential):
 def refresh_token(id: int, refresh_token: PostRefreshToken):
     return auth_repository.refresh(id, refresh_token)
 
+
+@app.delete("/auth/token", tags=["認証"], description="トークンを削除します ※ログアウト")
+def drop_token(token: str = Depends(get_auth_token)):
+    return auth_repository.remove(token)
+
 # ------------------------------------------------------------------------------
 
 
@@ -45,15 +66,21 @@ def get_items():
 
 
 @app.post("/item", tags=["アイテム"], description="新規登録します")
-def create_item(post_item: PostItem):
+def create_item(post_item: PostItem, isAuthorized: str = Depends(is_authorized)):
+    if not isAuthorized:
+        raise HTTPException(401)
     return item_repository.append(post_item)
 
 
 @app.put("/item/{id}", tags=["アイテム"], description="更新します")
-def update_item(id: int, post_item: PostItem):
+def update_item(id: int, post_item: PostItem, isAuthorized: str = Depends(is_authorized)):
+    if not isAuthorized:
+        raise HTTPException(401)
     return item_repository.update(id, post_item)
 
 
 @app.delete("/item/{id}", tags=["アイテム"], description="削除します")
-def remove_item(id: int):
+def remove_item(id: int, isAuthorized: str = Depends(is_authorized)):
+    if not isAuthorized:
+        raise HTTPException(401)
     return item_repository.remove(id)
